+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## 0.4.1 - 2026-01-15
+
+Addresses a gap in validation logic for instantiating the `CtsUrn` class.
+
+### Fixed
+
+- Added validation to ensure that neither `exemplar` nor `version` can be set when higher elements of the work hierarchy are `None`.
+
 ## 0.4.0 - 2026-01-15
 
 Non-breaking additions to the `CtsUrn` class.